requirements.txt
altgraph==0.17.3
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
inquirer==3.1.3
iterfzf==0.5.0.20.0
outcome==1.2.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
PySocks==1.7.1
python-editor==1.0.4
python-googlesearch==1.1.1
pyuseragents==1.0.5
readchar==4.0.5
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.3
wcwidth==0.2.6
wsproto==1.2.0
